# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "server API for Holy Hospital Application.",
        "title": "Holy Hospital Sever API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "ec2-3-91-177-221.compute-1.amazonaws.com",
    "basePath": "/api",
    "paths": {
...
        "/patient": {
            "get": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "description": "Fetch All Patient Data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patient"
                ],
                "summary": "GetAllPatient",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.MessageData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/response.Patient"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "417": {
                        "description": "Expectation Failed",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "description": "Fetch All Patient Data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patient"
                ],
                "summary": "CreatePatient",
                "parameters": [
                    {
                        "description": "Patient Details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Patient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.MessageOnly"
                        }
                    },
                    "417": {
                        "description": "Expectation Failed",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/patient/:id": {
            "get": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "description": "Fetch Patient Data By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patient"
                ],
                "summary": "GetPatientByID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.MessageData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.PatientDetails"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "417": {
                        "description": "Expectation Failed",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/patient/:id/delete": {
            "delete": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "description": "Fetch All Patient Data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patient"
                ],
                "summary": "DeletePatient",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.MessageOnly"
                        }
                    },
                    "417": {
                        "description": "Expectation Failed",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/patient/:id/update": {
            "put": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "description": "Fetch All Patient Data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patient"
                ],
                "summary": "UpdatePatient",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Patient Details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Patient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.MessageOnly"
                        }
                    },
                    "417": {
                        "description": "Expectation Failed",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.Patient": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "Sumenep"
                },
                "birthdate": {
                    "type": "string",
                    "example": "2001-04-14"
                },
                "blood_type": {
                    "type": "string",
                    "example": "A"
                },
                "code": {
                    "type": "string",
                    "example": "RM0001"
                },
                "full_name": {
                    "type": "string",
                    "example": "Faizur Ramadhan"
                },
                "gender": {
                    "type": "string",
                    "example": "Male"
                },
                "resident_registration": {
                    "type": "string",
                    "example": "8729301745162748"
                }
            }
        },
...
        "response.Error": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "response.MedicRecord": {
            "type": "object",
            "properties": {
                "complaint": {
                    "type": "string",
                    "example": "Sakit Perut"
                },
                "date_check": {
                    "type": "string",
                    "example": "2022-06-17"
                },
                "diagnose": {
                    "type": "string",
                    "example": "Maag"
                },
                "doctor": {
                    "type": "string",
                    "example": "Alsyad Ahmad"
                },
                "facility": {
                    "type": "string",
                    "example": "General"
                },
                "prescription": {
                    "type": "string",
                    "example": "Entrostop"
                },
                "serial_number": {
                    "type": "string",
                    "example": "RM/748/2022/0001"
                }
            }
        },
        "response.MessageData": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "response.MessageOnly": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "response.Patient": {
            "type": "object",
            "properties": {
                "birthdate": {
                    "type": "string",
                    "example": "2001-04-14"
                },
                "code": {
                    "type": "string",
                    "example": "RM0001"
                },
                "full_name": {
                    "type": "string",
                    "example": "Faizur Ramadhan"
                },
                "gender": {
                    "type": "string",
                    "example": "Male"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "resident_registration": {
                    "type": "string",
                    "example": "8729301745162748"
                }
            }
        },
        "response.PatientDetails": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "Sumenep"
                },
                "birthdate": {
                    "type": "string",
                    "example": "2001-04-14"
                },
                "blood_type": {
                    "type": "string",
                    "example": "A"
                },
                "code": {
                    "type": "string",
                    "example": "RM0001"
                },
                "full_name": {
                    "type": "string",
                    "example": "Faizur Ramadhan"
                },
                "gender": {
                    "type": "string",
                    "example": "Male"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "medic_record": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.MedicRecord"
                    }
                },
                "resident_registration": {
                    "type": "string",
                    "example": "8729301745162748"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKey": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}